requirements.txt
numpy==1.20.1
Pillow==8.2.0
Flask==2.0.1
Flask-Cors==3.0.10
waitress==2.0.0
